Introduce SoundArray type.
SoundArray objects will contain a set of sounds that can be played in a random
sequence. Added tests for methods to select the next sound, play and stop
sounds, and some related metadata methods such as size(). The tests for
ensuring the pseudo-randomness of choosing a sound are non-deterministic, and
should be fixed to guarantee that if the test passes it will pass every time.