Remove ionsound as a dependency.
Use HTML Audio objects directly to handle sound playback so that we can
observe events and update the UI accordingly. Add Vue as a dependency and
create a SoundEffect Vue component that handles playback of sounds and
displays feedback in the UI including icons for when the audio is playing
(using font-awesome) and keeping track of the play count.