Add a little more feedback to the UI.
Use requestAnimationFrame to animate the progress bar to give a smoother bar.
Display the name of the currently playing file in the sound effect's tile so
that it's a little easier to tell when different sounds are playing. Add a
case for the escape key in the keyup listener to cancel playback. Add some
event dispatch from the SoundEffect component so that one sound will interrupt
a playing sound.